update 7.1 afnetwork msdynamics pods

+++ TapShield/Utilities/TSUtilities.m
@@ -17,10 +17,10 @@
     NSInteger seconds = durationInSeconds % 60;
 
     if (hours > 0) {
-        return [NSString stringWithFormat:@"%02d:%02d:%02d", hours, minutes, seconds];
+        return [NSString stringWithFormat:@"%02ld:%02ld:%02ld", (long)hours, (long)minutes, (long)seconds];
     }
     else {
-        return [NSString stringWithFormat:@"%02d:%02d", minutes, seconds];
+        return [NSString stringWithFormat:@"%02ld:%02ld", (long)minutes, (long)seconds];
     }
 }
